requirements/prod.txt
#required for production of website.

asgiref==3.4.1
Babel==2.9.1
Django==3.2.9
django-phonenumber-field==6.0.0
phonenumbers==8.12.41
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pytz==2021.3
sqlparse==0.4.2